+++ examples/resourceMgmt.js
@@ -1,0 +1,74 @@
+/* Copyright (c) 2025, Oracle and/or its affiliates. */
+
+/******************************************************************************
+ *
+ * This software is dual-licensed to you under the Universal Permissive License
+ * (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl and Apache License
+ * 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose
+ * either license.
+ *
+ * If you elect to accept the software under the Apache License, Version 2.0,
+ * the following applies:
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * NAME
+ *   resourceMgmt.js
+ *
+ * DESCRIPTION
+ *   Example for using Explicit Resource Management feature of Node.js.
+ *   This feature is available from Node.js 24 onwards.
+ *   It uses resourceMgmtWorker.js, only if the Node.js version is 24
+ *   or later.
+ *   Check out the official Node.js 24 blog:
+ *   https://nodesource.com/blog/Node.js-version-24
+ *
+ *****************************************************************************/
+
+'use strict';
+
+Error.stackTraceLimit = 50;
+
+const oracledb = require('oracledb');
+
+// This example runs in both node-oracledb Thin and Thick modes.
+//
+// Optionally run in node-oracledb Thick mode
+if (process.env.NODE_ORACLEDB_DRIVER_MODE === 'thick') {
+
+  // Thick mode requires Oracle Client or Oracle Instant Client libraries.
+  // On Windows and macOS you can specify the directory containing the
+  // libraries at runtime or before Node.js starts.  On other platforms (where
+  // Oracle libraries are available) the system library search path must always
+  // include the Oracle library path before Node.js starts.  If the search path
+  // is not correct, you will get a DPI-1047 error.  See the node-oracledb
+  // installation documentation.
+  let clientOpts = {};
+  // On Windows and macOS platforms, set the environment variable
+  // NODE_ORACLEDB_CLIENT_LIB_DIR to the Oracle Client library path
+  if (process.platform === 'win32' || process.platform === 'darwin') {
+    clientOpts = { libDir: process.env.NODE_ORACLEDB_CLIENT_LIB_DIR };
+  }
+  oracledb.initOracleClient(clientOpts);  // enable node-oracledb Thick mode
+}
+
+console.log(oracledb.thin ? 'Running in thin mode' : 'Running in thick mode');
+const majorNodeJsVersion = parseInt(process.version.substring(1).split('.')[0]);
+
+if (majorNodeJsVersion < 24) {
+  console.log('This example requires Node.js 24 or later versions. Current Node.js version is', process.version);
+  process.exit();
+} else {
+  const resMgmtModule = require('./resourceMgmtWorker.js');
+  resMgmtModule.run().catch(console.error);
+}